# feat: implement binary node wasm