--- a/packages/baileys/src/Socket/messages-send.ts
+++ b/packages/baileys/src/Socket/messages-send.ts
@@ -25,7 +25,6 @@ import {
 	encryptMediaRetryRequest,
 	extractDeviceJids,
 	generateMessageIDV2,
-	generateParticipantHashV2,
 	generateWAMessage,
 	getStatusCodeForMediaRetry,
 	getUrlFromDirectPath,
@@ -931,10 +930,6 @@ export const makeMessagesSocket = (config: SocketConfig) => {
 				])
 				participants.push(...meNodes)
 				participants.push(...otherNodes)
-
-				if (meRecipients.length > 0 || otherRecipients.length > 0) {
-					extraAttrs['phash'] = generateParticipantHashV2([...meRecipients, ...otherRecipients])
-				}
 
 				shouldIncludeDeviceIdentity = shouldIncludeDeviceIdentity || s1 || s2
 			}