--- a/packages/baileys/src/WABinary/types.ts
+++ b/packages/baileys/src/WABinary/types.ts
@@ -1,4 +1,3 @@
-import * as constants from './constants'
 /**
  * the binary node WA uses internally for communication
  *
@@ -13,5 +12,3 @@ export type BinaryNode = {
 }
 export type BinaryNodeAttributes = BinaryNode['attrs']
 export type BinaryNodeData = BinaryNode['content']
-
-export type BinaryNodeCodingOptions = typeof constants