--- a/packages/whatsapp-rust-bridge/test/flat.test.ts
+++ b/packages/whatsapp-rust-bridge/test/flat.test.ts
@@ -444,3 +444,4 @@ describe("flat codec intern table", () => {
     expect(poolOf(flat)).toEqual(["k", repeated]);
   });
 });
+