--- a/packages/whatsapp-rust-bridge/ts/index.ts
+++ b/packages/whatsapp-rust-bridge/ts/index.ts
@@ -11,3 +11,4 @@ function resolveWasm(variant: WasmVariant): URL {
 
 export const __wasmSimdActive: boolean = initializeWasm(resolveWasm);
 export * from "../pkg/whatsapp_rust_bridge.js";
+export * from "./flat.js";