… packages/baileys/src/WABinary/constants.ts
… packages/baileys/src/__tests__/Utils/generics.test.ts
… packages/baileys/src/__tests__/binary/wabinary-codec.test.ts
… packages/whatsapp-rust-bridge/Cargo.lock
… packages/whatsapp-rust-bridge/benches/binary-tree.ts
… packages/whatsapp-rust-bridge/package.json
… packages/whatsapp-rust-bridge/scripts/check-package.mjs
… packages/whatsapp-rust-bridge/src/binary.rs
… packages/whatsapp-rust-bridge/test/binary.test.ts
… packages/whatsapp-rust-bridge/test/handshake-parity.test.ts
… packages/whatsapp-rust-bridge/test/parity.test.ts
… packages/whatsapp-rust-bridge/test/server-response-parity.test.ts
… packages/whatsapp-rust-bridge/ts/flat.ts
… packages/whatsapp-rust-bridge/ts/wasm-runtime.ts