This pull request introduces a major refactor in how binary nodes are encoded and decoded, shifting from custom TypeScript implementations to using a high-performance Rust-based library (`whatsapp-rust-bridge`). It also standardizes the use of `Buffer` and `Uint8Array` types across the codebase for better interoperability and reliability. Additionally, several minor bug fixes and improvements are made to ensure correct type handling and event emission.

**Migration to Rust-based binary node encoding/decoding:**

* Replaced custom implementations in `src/WABinary/decode.ts` and `src/WABinary/encode.ts` with calls to the Rust-based `whatsapp-rust-bridge` library, removing large amounts of legacy code and improving performance and maintainability. (`src/WABinary/decode.ts`, [[1]](diffhunk://#diff-3559666031be8b9a8d861b29c393dca515af074c6031657525c7907f7fa93245L1-R10); `src/WABinary/encode.ts`, [[2]](diffhunk://#diff-8c7760a4f598f47b81900ae903442c913273d85208b0a12c7f21e4511a513087L1-R7)
* Added the `whatsapp-rust-bridge` dependency to `package.json`.

**Standardization of Buffer/Uint8Array usage:**

* Updated types in `src/Types/Signal.ts` to consistently use `Buffer` instead of `Uint8Array` for cryptographic and binary message fields, and updated function signatures accordingly. [[1]](diffhunk://#diff-782c7555da6cfdccbf49d934d85968ae45a1dce63ddc6c9bc2ace03f245ec7baL6-R6) [[2]](diffhunk://#diff-782c7555da6cfdccbf49d934d85968ae45a1dce63ddc6c9bc2ace03f245ec7baL17-R22) [[3]](diffhunk://#diff-782c7555da6cfdccbf49d934d85968ae45a1dce63ddc6c9bc2ace03f245ec7baL33-R42) [[4]](diffhunk://#diff-782c7555da6cfdccbf49d934d85968ae45a1dce63ddc6c9bc2ace03f245ec7baL55-R58)
* Updated utility functions and message processing logic to convert between `Uint8Array` and `Buffer` as needed, ensuring compatibility with new encoding/decoding logic. (`src/WABinary/generic-utils.ts`, [[1]](diffhunk://#diff-f1d2069de7e0d015176d0c0a6526fd074823f7b6667fb8680c5aea416b6bd211L29-R32); `src/Utils/decode-wa-message.ts`, [[2]](diffhunk://#diff-c20f2d846a8b973ca5d51a65f300f12bd449812f57b12481879c1887d471927aL184-R184) [[3]](diffhunk://#diff-c20f2d846a8b973ca5d51a65f300f12bd449812f57b12481879c1887d471927aL193-R193); `src/Utils/signal.ts`, [[4]](diffhunk://#diff-66d5215ceb1308a4f1649d0c22ff4dc36a6bb5826a05ff6389314a07de4ea24fL93-R93) [[5]](diffhunk://#diff-66d5215ceb1308a4f1649d0c22ff4dc36a6bb5826a05ff6389314a07de4ea24fL122-R122); `src/Utils/chat-utils.ts`, [[6]](diffhunk://#diff-b71adcec0c08ff3a9ee4e1eae36ec1afba918134ebc17fd41bc5a80bcf674054L337-R337); `src/Utils/validate-connection.ts`, [[7]](diffhunk://#diff-f8ae433f18edd8479d0734d7777bf2c482cff8883686251e68373f0da26c7685L137-R139)

**Bug fixes and event handling improvements:**

* Fixed event emission and type checking in socket-related files to use `instanceof Uint8Array` for binary data, preventing unexpected runtime errors. (`src/Socket/messages-recv.ts`, [[1]](diffhunk://#diff-d46b41f0350807458fc3f16bec3e0e06908854d34cb7c940d63f25f8bdbcef65L401-R401); `src/Socket/socket.ts`, [[2]](diffhunk://#diff-2f6a136112cda2ac0303f24b4c09666617d7926db608b4483a310c9104127825L307-R306) [[3]](diffhunk://#diff-2f6a136112cda2ac0303f24b4c09666617d7926db608b4483a310c9104127825L598-R602)
* Updated QR code generation and message sending logic to handle new binary types and asynchronous encoding. (`src/Socket/socket.ts`, [[1]](diffhunk://#diff-2f6a136112cda2ac0303f24b4c09666617d7926db608b4483a310c9104127825L132-R136) [[2]](diffhunk://#diff-2f6a136112cda2ac0303f24b4c09666617d7926db608b4483a310c9104127825L598-R602)

**General code cleanup:**

* Removed unused imports and legacy code from files affected by the migration to the Rust bridge, making the codebase more maintainable. (`src/WABinary/decode.ts`, [[1]](diffhunk://#diff-3559666031be8b9a8d861b29c393dca515af074c6031657525c7907f7fa93245L1-R10); `src/WABinary/encode.ts`, [[2]](diffhunk://#diff-8c7760a4f598f47b81900ae903442c913273d85208b0a12c7f21e4511a513087L1-R7)
* Ensured all binary node handling is now centralized through the new Rust bridge APIs, reducing duplication and complexity. [[1]](diffhunk://#diff-3559666031be8b9a8d861b29c393dca515af074c6031657525c7907f7fa93245L1-R10) [[2]](diffhunk://#diff-8c7760a4f598f47b81900ae903442c913273d85208b0a12c7f21e4511a513087L1-R7)

**Socket and message pipeline adjustments:**

* Made the `sendNode` function asynchronous to accommodate the new async encoding logic. (`src/Socket/socket.ts`, [src/Socket/socket.tsL132-R136](diffhunk://#diff-2f6a136112cda2ac0303f24b4c09666617d7926db608b4483a310c9104127825L132-R136))
* Minor fix to remove an unnecessary blank line in `src/Socket/socket.ts`.

This refactor sets the foundation for improved performance and reliability in handling WhatsApp binary nodes, while also simplifying future maintenance and feature development.

The whatsapp rust bridge (wasm stuffs) repository: https://github.com/jlucaso1/whatsapp-rust-bridge
The whatsapp rust: https://github.com/jlucaso1/whatsapp-rust

This is totally compatible with web environments too and has a great margin to optimize the bridge code (zero copy, improve serialization between JS and RUST, etc)