@purpshell 2025-08-19
Can you benchmark this? Mock test sendNode and pass through a thousand or so nodes and also make sure to make some different nodes chock-full of data and fields and test the difference.