@jlucaso1 2025-08-20
[bench.zip](https://github.com/user-attachments/files/21877729/bench.zip)
Benchmark results:
```bash
#using deprecated parameters for the initialization function; pass a single object instead # need to solve this warning later
clk: ~4.21 GHz
cpu: 12th Gen Intel(R) Core(TM) i5-12450H
runtime: node 22.18.0 (x64-linux)

benchmark                   avg (min … max) p75 / p99    (min … top 1%)
------------------------------------------- -------------------------------
• Encoding (JS Object -> Binary)
------------------------------------------- -------------------------------
Rust WASM (warmup-ignore)     45.61 µs/iter  50.37 µs  ▅█                  
                     (34.73 µs … 119.46 µs)  70.90 µs ▄███▅██▅▃▇▄▃▃▂▃▂▃▃▁▁▁
                  gc(  2.15 ms …   4.51 ms)   3.59 kb (  3.11 kb… 26.16 kb)
                   0.36 ipc ( 46.18% cache)  740.92 branch misses
        129.58k cycles  46.62k instructions   4.70k c-refs   2.53k c-misses

Rust WASM (small)              5.25 µs/iter   5.34 µs █▂▂█   ▆▂            
                        (4.97 µs … 5.92 µs)   5.90 µs ████▃▅▅██▇▁▅▁▁▃▃▃▃▁▁▃
                  gc(  2.24 ms …   5.43 ms)   3.11 kb (  3.09 kb…  3.52 kb)
                   3.27 ipc ( 69.45% cache)   17.71 branch misses
         18.94k cycles  61.92k instructions   67.10 c-refs   20.50 c-misses

Rust WASM (medium)             8.77 µs/iter   9.44 µs  █    ▅              
                       (7.14 µs … 12.38 µs)  11.45 µs ▇█▄▄▄▁█▇▇▇▄▇▄▄▄▁▁▄▁▁▄
                  gc(  2.41 ms …   6.44 ms) 218.68  b (214.03  b…364.67  b)
                   3.30 ipc ( 64.43% cache)   23.56 branch misses
         27.61k cycles  91.22k instructions   97.38 c-refs   34.64 c-misses

Rust WASM (large)            184.65 µs/iter 219.50 µs  ▇█▃                 
                    (128.17 µs … 358.25 µs) 332.92 µs ▃███▆▆▂▃▄▄▃▅▃▂▂▄▂▂▁▁▁
                  gc(  3.04 ms …   8.93 ms)  36.75 kb ( 29.88 kb… 77.92 kb)
                   1.75 ipc ( 35.92% cache)   1.68k branch misses
        397.69k cycles 695.75k instructions   6.54k c-refs   4.19k c-misses

Old js (warmup-ignore)        66.52 µs/iter  71.58 µs  ▃██▃▂▃              
                     (46.67 µs … 178.23 µs) 126.36 µs ▃██████▆▄▅▃▂▁▂▂▁▂▁▁▁▁
                  gc(  3.00 ms …   6.82 ms)   9.19 kb (  8.58 kb… 56.84 kb)
                   0.52 ipc ( 43.39% cache)  941.18 branch misses
        175.45k cycles  91.09k instructions   6.32k c-refs   3.58k c-misses

Old js (small)                14.14 µs/iter  14.43 µs   ▃     ▃ █   ▃      
                      (13.26 µs … 15.20 µs)  15.07 µs ▆▁█▆▁▆▁▆█▆█▆▁▆█▆▁▁▁▁▆
                  gc(  2.37 ms …   3.62 ms)   2.88 kb (  2.84 kb…  3.47 kb)
                   3.58 ipc ( 57.80% cache)   32.29 branch misses
         48.83k cycles 174.80k instructions  391.15 c-refs  165.06 c-misses

Old js (medium)               20.62 µs/iter  20.71 µs             █     █  
                      (20.34 µs … 20.77 µs)  20.75 µs █▁▁▁▁▁▁█▁▁▁███▁██▁█▁█
                  gc(  2.49 ms …   4.26 ms)   2.59 kb (  2.53 kb…  2.74 kb)
                   3.68 ipc ( 57.36% cache)   48.53 branch misses
         71.93k cycles 264.83k instructions  622.16 c-refs  265.31 c-misses

Old js (large)               371.65 µs/iter 401.43 µs   ▅█▃▄ ▇             
                    (264.36 µs … 750.28 µs) 665.32 µs ▄▇██████▆▆▂▂▁▂▂▁▂▁▁▁▂
                  gc(  2.11 ms …   9.01 ms) 358.00 kb (  2.30 kb…944.30 kb)
                   2.62 ipc ( 53.75% cache)   3.10k branch misses
          1.18M cycles   3.08M instructions  16.05k c-refs   7.42k c-misses
```