@devlikepro 2025-08-20
So rust bridge is only about **2x** faster than the original JS version, right?
Not sure if adding Rust to the pure-js-typescript ecosystem is really worth it 🤔

Libsignal, for example, eats way more CPU time anyway.