@purpshell 2025-08-20
> So rust bridge is only about **2x** faster than the original JS version, right? Not sure if adding Rust to the pure-js-typescript ecosystem is really worth it 🤔
> 
> Libsignal, for example, eats way more CPU time anyway.

We are planning to move libsignal under this rust bridge, and WASM allows for all runtime support. Also, the rust bridge is still *not optimized*. I see a promising future for this