@vinikjkkj 2025-08-29
> i rewrited this using pure js and it's now x10 faster...

share with us