@Santosl2 2025-09-01
> i rewrited this using pure js and it's now x10 faster...

share with us