@jlucaso1 2025-09-01
I've made some optimizations (reducing allocations and removed some unecessary operations in the rust code).

Improved execution time and memory usage in operations.

New benchmark results ([bench.zip](https://github.com/user-attachments/files/22070877/bench.zip)):
```bash
➜  Baileys git:(feat-binary-node-rust-wasm) ✗ yarn tsx --expose-gc bench/bench.ts
clk: ~4.18 GHz
cpu: 12th Gen Intel(R) Core(TM) i5-12450H
runtime: node 22.18.0 (x64-linux)

benchmark                   avg (min … max) p75 / p99    (min … top 1%)
------------------------------------------- -------------------------------
• Encoding (JS Object -> Binary)
------------------------------------------- -------------------------------
Rust WASM (small)              4.92 µs/iter   4.95 µs          ▄█▅▄        
                        (4.68 µs … 5.17 µs)   5.14 µs ▃▁▁▁▃▆▄▆▄████▆▃▁▄▃▁▁▃
                  gc(  2.17 ms …   3.48 ms)   2.47 kb (  2.47 kb…  2.60 kb)
                   3.36 ipc ( 74.49% cache)   16.01 branch misses
         18.56k cycles  62.41k instructions   48.50 c-refs   12.37 c-misses

Rust WASM (medium)             7.28 µs/iter   7.31 µs    ▄ █     ▄▄ ▄█     
                        (7.20 µs … 7.36 µs)   7.35 µs ▅▅▅█▁█▅█▅█▅██▅██▁█▅▅▅
                  gc(  2.21 ms …   3.19 ms)   3.52 kb (  3.52 kb…  3.52 kb)
                   3.47 ipc ( 74.73% cache)   22.98 branch misses
         27.23k cycles  94.37k instructions   66.81 c-refs   16.89 c-misses

Rust WASM (large)             59.83 µs/iter  59.93 µs █       █      █     
                      (59.51 µs … 60.24 µs)  60.07 µs ██▁▁▁▁▁▁█▁██▁▁▁█▁▁▁██
                  gc(  2.28 ms …   3.21 ms)   1.01 kb (  1.01 kb…  1.01 kb)
                   3.61 ipc ( 75.05% cache)  254.49 branch misses
        222.48k cycles 802.85k instructions  542.33 c-refs  135.31 c-misses

Old js (small)                88.79 µs/iter  90.01 µs  █▇                  
                     (76.21 µs … 158.24 µs) 140.38 µs ▅███▅▃▁▂▂▃▄▃▂▂▂▁▁▁▁▁▁
                  gc(  2.10 ms …   4.12 ms)  27.64 kb ( 22.95 kb…119.06 kb)
                   1.14 ipc ( 56.55% cache)   1.53k branch misses
        271.81k cycles 309.33k instructions   7.89k c-refs   3.43k c-misses

Old js (medium)               19.43 µs/iter  19.46 µs             █        
                      (19.26 µs … 19.60 µs)  19.59 µs █▁▁▅▅▁▁▁▁▅▅▁█▁▁▁▁▅▁▁▅
                  gc(  2.39 ms …   4.97 ms)   1.96 kb (  1.88 kb…  2.21 kb)
                   3.68 ipc ( 59.53% cache)   56.31 branch misses
         71.58k cycles 263.50k instructions  607.44 c-refs  245.82 c-misses

Old js (large)               352.57 µs/iter 389.67 µs   ▂▃ ▆█▃ ▇▄▄▆▂▄ ▂▅   
                    (258.33 µs … 520.43 µs) 447.75 µs ▄▃██▅██████████▅██▆▄▄
                  gc(  2.32 ms …   8.13 ms) 391.87 kb ( 54.12 kb…  1.22 mb)
                   2.67 ipc ( 55.72% cache)   3.18k branch misses
          1.19M cycles   3.18M instructions  15.88k c-refs   7.03k c-misses
```