@AstroX11 2025-09-22
> i rewrited this using pure js and it's now x10 faster...

vague claims.